refactor: extract the mapstats as customhooks
